แก้ไขเป็น uint64

+++ controllers/auth_controller.go
@@ -298,7 +298,7 @@ func VerifyUser(c *gin.Context) {
 		return
 	}
 
-	roleID := uint(4)
+	roleID := uint64(4)
 	user := models.User{
 		Gmail:        userVerif.Gmail,
 		Password:     userVerif.Password,
